requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.20
ragas>=0.1.0
boto3>=1.34.0
botocore>=1.34.0

# AWS and CloudWatch
watchtower>=3.0.0
aws-lambda-powertools>=2.30.0

# Data processing
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
moto>=5.0.0

# CLI and utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Optional: For architecture diagrams
diagrams>=0.23.0
